+++ LDR/LDR/ViewModels/MainInterfaceViewModel.swift
@@ -12,7 +12,6 @@ import Foundation
 @MainActor
 class MainInterfaceViewModel: ObservableObject {
     @Published var currentUserId: String = ""
-    @Published var ifExist = false
     private var handler: AuthStateDidChangeListenerHandle?
     
     init() {
@@ -21,26 +20,13 @@ class MainInterfaceViewModel: ObservableObject {
                 self?.currentUserId = user?.uid ?? ""
             }
         }
-        isUserExist()
     }
     
     var isSignIn: Bool {
         return Auth.auth().currentUser != nil
     }
     
-    func isUserExist() {
-        let currentUserId = Auth.auth().currentUser?.uid ?? "a"
-        let db = Firestore.firestore()
-        //var ifExist = false
-        let userRef = db.collection("users").document(currentUserId)
-        userRef.getDocument { (document, error) in
-            if let document = document, document.exists {
-                self.ifExist = true
-                //return
-            } else {
-                //return
-            }
-        }
-        //return ifExist
+    var isEmailSetUp: Bool {
+        return Auth.auth().currentUser?.email != nil
     }
 }
