+++ LDR/LDR/Views/LoginView.swift
@@ -16,6 +16,8 @@ struct LoginView: View {
     @Bindable private var login = LoginViewViewModel()
     @FocusState private var focusField: Field?
     
+    @EnvironmentObject var signInStatus: SignInEnvironment
+    
     enum Field {
         case email, password
     }
@@ -42,13 +44,6 @@ struct LoginView: View {
             
             Spacer()
             
-            // Create Account
-            VStack {
-                Text("New around here?")
-                NavigationLink("Create An Account",
-                               destination: RegisterView())
-            }
-            .padding()
         }
     }
     
@@ -89,6 +84,7 @@ struct LoginView: View {
             login.appleLoginRequest(request)
         } onCompletion: { result in
             login.appleLoginCompletion(result)
+            signInStatus.isUserExist()
         }
         .signInWithAppleButtonStyle(.black)
         .clipShape(RoundedRectangle(cornerRadius: 25.0))
@@ -123,6 +119,8 @@ struct LoginView: View {
                 Auth.auth().signIn(with: credential) {_,error in
                     
                 }
+                
+                signInStatus.isUserExist()
                 
                 let emailAddress = user.profile?.email
